migraciones

+++ site/database/models/actorEpisode.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
     const ActorEpisode = sequelize.define(
         "ActorEpisode",
         {
-            screentime: DataTypes.INTEGER(10),
+            screentime: DataTypes.FLOAT(3,1),
         },
         {
             tableName: "actorepisode",
